Host field too small, increased size GET
The host field contains most of the time URLs, therefore the field
should be able to contain urls with a length > 100 char, because
browsers and webservers limit URLs to 2KB (this may vary).
In a custom provider I'm working on, I use the host field to store even
more data about this host in json, and this exceeds 100 chars as well.